fixed the minor flaw of data type transition

+++ WeAlumni/DataEncrptionInterface.cpp
@@ -17,15 +17,14 @@
 using namespace System;
 
 /*
- * Encrypt
- * This method will encrypt the given plain text into the cipher text by algorithm.
- * @param String^ plain text that will be encrypted
- * @return String^ cipher text that is encrpyted
+ * MarshalString
+ * This method will convert a System::String^ to std::string.
+ * @param System::String^ string needs to be converted
+ * @return std::string string converted
  */
 std::string MarshalString(String^ s) {
     using namespace Runtime::InteropServices;
-    const char* chars =
-        (const char*)(Marshal::StringToHGlobalAnsi(s)).ToPointer();
+    const char* chars = (const char*)(Marshal::StringToHGlobalAnsi(s)).ToPointer();
     std::string os = chars;
     Marshal::FreeHGlobal(IntPtr((void*)chars));
     return os;
@@ -40,5 +39,5 @@ std::string MarshalString(String^ s) {
 int WeAlumni::DataEncryption::Encrypt(String^ plain) {
     std::string plainText = MarshalString(plain);
     std::hash<std::string> str_hash;
-    return str_hash(plainText);
+    return (int)str_hash(plainText);
 }
